requirements.txt
Flask==3.0.3
PyJWT==2.8.0
bcrypt==4.1.3
psycopg2-binary>=2.9.9
supabase==2.4.6
python-dotenv==1.0.1
Flask-Limiter==3.7.0
email-validator==2.1.1
pytest==8.2.2
pytest-flask==1.3.0
requests==2.32.3
